Got AI assist working again
