+++ src/client/CodeEditor/getOrCreateEditor.ts
@@ -34,13 +34,8 @@ import {
   EditorCacheValue,
 } from '../useEditorCache';
 import { ThemeLabel, themeOptionsByLabel } from '../themes';
-import {
-  AIAssistCodeMirrorKeyMap,
-  AIAssistState,
-} from '../AIAssist';
 import { typeScriptCompletions } from './typeScriptCompletions';
 import { typeScriptLinter } from './typeScriptLinter';
-import { TabState } from '../vzReducer';
 import { keymap } from '@codemirror/view';
 import { basicSetup } from './basicSetup';
 
@@ -90,11 +85,7 @@ export const getOrCreateEditor = ({
   onInteract,
   editorCache,
   usernameRef,
-  aiAssistEndpoint,
-  aiAssistOptions,
   typeScriptWorker,
-  tabList,
-  aiAssistStateRef,
 }: {
   fileId: FileId;
 
@@ -115,12 +106,7 @@ export const getOrCreateEditor = ({
   editorCache: EditorCache;
   usernameRef: React.MutableRefObject<Username>;
   aiAssistEndpoint?: string;
-  aiAssistOptions?: {
-    [key: string]: any;
-  };
   typeScriptWorker: Worker;
-  tabList: Array<TabState>;
-  aiAssistStateRef: React.MutableRefObject<AIAssistState>;
 }): EditorCacheValue => {
   // Cache hit
   if (editorCache.has(fileId)) {
@@ -236,16 +222,15 @@ export const getOrCreateEditor = ({
 
   extensions.push(rotationIndicator);
 
-  extensions.push(
-    AIAssistCodeMirrorKeyMap({
-      shareDBDoc,
-      fileId,
-      tabList,
-      aiAssistEndpoint,
-      aiAssistOptions,
-      aiAssistStateRef,
-    }),
-  );
+  // extensions.push(
+  //   AIAssistCodeMirrorKeyMap({
+  //     shareDBDoc,
+  //     fileId,
+  //     tabList,
+  //     aiAssistEndpoint,
+  //     aiAssistOptions,
+  //   }),
+  // );
 
   // Add the extension that provides TypeScript completions.
   if (enableTypeScriptCompletions) {
