+++ test/sampleDirectories/kitchenSink/index.js
@@ -1,3 +1,3 @@
 const n = 5;
 // Add 20 to this number
-const 
+const newNum =
